Docker-ise and Spring-boot-ify
Starts the process of packaging Speedith into a Docker container, and
wrapping it in a ReMix plug-in server. The docker-ising isolates the
old and insecure Java 7 from the rest of the world, even if Docker is
not meant to be a build environment.

NOTE: the directory 'local_m2' is missing from the commit. It contains
all the maven dependencies, so that the build environment does not
have to download them again every time the container is
created. Including it is wasteful, particularly given it can be
extracted easily enough from the container on the first run.