Top Level, Reasoning Control, Success, ReMixFormula
The top level controller is active, providing the plug-in JSON object
to the ReMix system correctly. So far, only the reasoning extension is
valid.

The reasoning controller has been started, exposing a list of possible
rules. The rules do not actually work yet.

A few utility classes have also been defined. This includes Success, a
surprisingly complicated class to produce the correct JSON in the
response. There is also a new ReMixFormula class, which wraps around
the content received from ReMix (and expected back by ReMix).